--- .hgignore
@@ -1,17 +1,0 @@
-relre:^_ReSharper.*
-relre:.*obj/.*
-relre:.*bin/.*
-relre:.*\.user$
-relre:^_UpgradeReport_Files.*
-relre:.*\.suo$
-glob:UpgradeLog.XML
-glob:*.vs10x
-glob:nuget/dotnetdbf/lib/Net20/*
-glob:nuget/dotnetdbf/lib/Net40/*
-glob:nuget/NuGet.exe
-glob:nuget/*.nupkg
-glob:dist/*
-glob:nuget/dotnetdbf.enumerable/lib/Net40/*
-syntax: glob
-packages/*
-*.swp
