--- .hgtags
@@ -1,8 +1,0 @@
-43c37dd96ef92f9e03897f7d11e917bf9e6fe7f0 1_0_0
-85feb1f9254f8f3ef63cd5fac2fc696e8a53a598 1_0_1
-2921ee3800fa3fb192d17af659c49c2fc865f1d0 1_0_2
-6615b74c5f7845f7e7a8dd1574943d9b52f8ccc7 3_0_1
-55b6c37a1cbfdbf501f4b1cc8c52dcf5dca4d2e2 3_0_2
-55b6c37a1cbfdbf501f4b1cc8c52dcf5dca4d2e2 3_0_3
-b3a4bcdb9953b1d5dadfc46ca78d11b5307ba00c 3.0.4
-aeb9a414d6e18aa032bb887b07814a00a89bba14 3.1.0/4.1.0
