--- DotNetDBF.Enumerable/packages.config
@@ -1,4 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="ImpromptuInterface" version="6.1.3" targetFramework="net40" />
-</packages>
