--- DotNetDBF/app.config
@@ -1,3 +1,0 @@
-<?xml version="1.0"?>
-<configuration>
-	<startup/></configuration>
