--- nuget/dotnetdbf.enumerable.nuspec
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?> 
-<package> 
-  <metadata> 
-    <id>dotnetdbf.enumerable</id> 
-    <version>4.3.10.4</version> 
-    <authors>Jay Tuley</authors> 
-    <owners>Jay Tuley</owners>
-    <description>For dotnetdbf projects using .net 4.0 projects this is an enumeration framework in which makes it easy to use Linq to Objects.</description> 
-    <projectUrl>http://code.google.com/p/dotnetdbf/</projectUrl>
-    <licenseUrl>http://www.gnu.org/licenses/lgpl.html</licenseUrl>
-    <tags>clipper dbf .net40</tags>
-    <language>en-US</language> 
-    <dependencies>
-	<dependency id="dotnetdbf" version="3.1.10" />
- 	<dependency id="ImpromptuInterface" version="6.1.3" />
-    </dependencies>
-  </metadata>
-  <files>
-    <file src="..\COPYING.TXT"/>
-    <file src="..\DotNetDBF.Enumerable\bin\Release\DotNetDBF.Enumerable.dll" target="lib\net40"/>
-    <file src="..\DotNetDBF.Enumerable\bin\Release\DotNetDBF.Enumerable.xml" target="lib\net40"/>
-  </files>
-</package>
-
