--- nuget/dotnetdbf.nuspec
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?> 
-<package> 
-  <metadata> 
-    <id>dotnetdbf</id> 
-    <version>3.1.10</version> 
-    <authors>Anil Kumar, Jay Tuley</authors> 
-    <owners>Jay Tuley</owners>
-    <summary>This is a basic file parser for reading and writing xBase DBF files, particularly Clipper.</summary>
-    <description>This is a basic file parser for reading and writing xBase DBF files particularlly Clipper. Code derived from javadbf. </description> 
-    <projectUrl>http://code.google.com/p/dotnetdbf/</projectUrl>
-    <licenseUrl>http://www.gnu.org/licenses/lgpl.html</licenseUrl>
-    <tags>clipper dbf .net35</tags>
-    <language>en-US</language>
-  </metadata>
-  <files>
-    <file src="..\COPYING.TXT"/>
-    <file src="..\dotnetdbf\bin\Release\*.dll" target="lib\net35"/>
-    <file src="..\dotnetdbf\bin\Release\*.pdb" target="lib\net35"/>
-    <file src="..\dotnetdbf\bin\Release\*.xml" target="lib\net35"/>
-  </files>
-</package>
-
